backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
dnspython>=2.6.1
cryptography>=43.0.0
anthropic>=0.34.0
openai>=1.40.0
twilio>=9.0.0
aiosqlite>=0.20.0
rich>=13.7.0
python-multipart>=0.0.9

# --- P0 upgrades ---
scikit-learn>=1.5.0          # IsolationForest behavioural anomaly detection
numpy>=1.26.0                # feature vectors
reportlab>=4.2.0             # PDF CISO board report

# --- P1 upgrades ---
sse-starlette>=2.1.0         # server-sent events for live dashboard
pypdf>=4.3.0                 # DPDP Act PDF loader (optional)

# --- dev / tests ---
pytest>=8.3.0
pytest-asyncio>=0.23.0
